multi-agents/requirements.txt
crewai>=0.1.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
python-dotenv>=0.19.0
langchain>=0.1.0
google-generativeai>=0.3.0
pymongo>=4.0.0
qdrant-client>=1.7.0
